+++ examples/decode-from-separate-data-using-loader.ts
@@ -15,7 +15,7 @@ const renderer = new WebGLRenderer()
 
 const loader = new GainMapLoader(renderer)
 
-const result = loader.load(['sdr.jpeg', 'gainmap.jpeg', 'metadata.json'])
+const result = await loader.loadAsync(['sdr.jpeg', 'gainmap.jpeg', 'metadata.json'])
 // `result` can be used to populate a Texture
 
 const scene = new Scene()
@@ -34,3 +34,7 @@ renderer.render(scene, new PerspectiveCamera())
 scene.background = result.toDataTexture()
 scene.background.mapping = EquirectangularReflectionMapping
 scene.background.minFilter = LinearFilter
+
+// result must be manually disposed
+// when you are done using it
+result.dispose()