+++ src/decode/loaders/LoaderBase.ts
@@ -15,26 +15,16 @@ import {
 } from 'three'
 
 import { QuadRenderer } from '../../core/QuadRenderer'
-import { type GainMapMetadata } from '../../core/types'
+import { type GainMapMetadata, QuadRendererTextureOptions } from '../../core/types'
 import { GainMapDecoderMaterial } from '../materials/GainMapDecoderMaterial'
-
-/**
- * private function, async get image from blob
- *
- * @param blob
- * @returns
- */
-const getImage = (blob: Blob) => {
-  return new Promise<HTMLImageElement>((resolve, reject) => {
-    const img = document.createElement('img')
-    img.onload = () => { resolve(img) }
-    img.onerror = (e) => { reject(e) }
-    img.src = URL.createObjectURL(blob)
-  })
-}
+import { getHTMLImageFromBlob } from '../utils/get-html-image-from-blob'
 
 export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFloatType, GainMapDecoderMaterial>, TUrl> {
   private _renderer: WebGLRenderer
+  private _renderTargetOptions?: QuadRendererTextureOptions
+  /**
+   * @private
+   */
   protected _internalLoadingManager: LoadingManager
   /**
    *
@@ -45,6 +35,17 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
     super(manager)
     this._renderer = renderer
     this._internalLoadingManager = new LoadingManager()
+  }
+
+  /**
+   * Specify the renderTarget options to use when rendering the gain map
+   *
+   * @param options
+   * @returns
+   */
+  public setRenderTargetOptions (options: QuadRendererTextureOptions) {
+    this._renderTargetOptions = options
+    return this
   }
 
   /**
@@ -66,38 +67,39 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
       sdr: new Texture()
     })
 
-    return new QuadRenderer(
-      16,
-      16,
-      HalfFloatType,
-      LinearSRGBColorSpace,
+    return new QuadRenderer({
+      width: 16,
+      height: 16,
+      type: HalfFloatType,
+      colorSpace: LinearSRGBColorSpace,
       material,
-      this._renderer
-    )
+      renderer: this._renderer,
+      renderTargetOptions: this._renderTargetOptions
+    })
   }
 
   /**
-   *
-   * @private
-   * @param quadRenderer
-   * @param gainMapBuffer
-   * @param sdrBuffer
-   * @param metadata
-   */
-  protected async render (quadRenderer: QuadRenderer<typeof HalfFloatType, GainMapDecoderMaterial>, gainMapBuffer: ArrayBuffer | string, sdrBuffer: ArrayBuffer | string, metadata: GainMapMetadata) {
-    const gainMapBlob = new Blob([gainMapBuffer], { type: 'image/jpeg' })
+ * @private
+ * @param quadRenderer
+ * @param metadata
+ * @param sdrBuffer
+ * @param gainMapBuffer
+ */
+  protected async render (quadRenderer: QuadRenderer<typeof HalfFloatType, GainMapDecoderMaterial>, metadata: GainMapMetadata, sdrBuffer: ArrayBuffer, gainMapBuffer?: ArrayBuffer) {
+    // this is optional, will render a black gain-map if not present
+    const gainMapBlob = gainMapBuffer ? new Blob([gainMapBuffer], { type: 'image/jpeg' }) : undefined
 
     const sdrBlob = new Blob([sdrBuffer], { type: 'image/jpeg' })
 
     let sdrImage: ImageBitmap | HTMLImageElement
-    let gainMapImage: ImageBitmap | HTMLImageElement
+    let gainMapImage: ImageBitmap | HTMLImageElement | undefined
 
     let needsFlip = false
 
     if (typeof createImageBitmap === 'undefined') {
       const res = await Promise.all([
-        getImage(gainMapBlob),
-        getImage(sdrBlob)
+        gainMapBlob ? getHTMLImageFromBlob(gainMapBlob) : Promise.resolve(undefined),
+        getHTMLImageFromBlob(sdrBlob)
       ])
 
       gainMapImage = res[0]
@@ -106,7 +108,7 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
       needsFlip = true
     } else {
       const res = await Promise.all([
-        createImageBitmap(gainMapBlob, { imageOrientation: 'flipY' }),
+        gainMapBlob ? createImageBitmap(gainMapBlob, { imageOrientation: 'flipY' }) : Promise.resolve(undefined),
         createImageBitmap(sdrBlob, { imageOrientation: 'flipY' })
       ])
 
@@ -114,7 +116,7 @@ export class LoaderBase<TUrl = string> extends Loader<QuadRenderer<typeof HalfFl
       sdrImage = res[1]
     }
 
-    const gainMap = new Texture(gainMapImage,
+    const gainMap = new Texture(gainMapImage || new ImageData(2, 2),
       UVMapping,
       ClampToEdgeWrapping,
       ClampToEdgeWrapping,