+++ src/encode/get-gainmap.ts
@@ -25,11 +25,19 @@ export const getGainMap = (params: { sdr: InstanceType<typeof QuadRenderer> } & 
     hdr: dataTexture
   })
 
-  const quadRenderer = new QuadRenderer(dataTexture.image.width, dataTexture.image.height, UnsignedByteType, LinearSRGBColorSpace, material, renderer)
+  const quadRenderer = new QuadRenderer({
+    width: dataTexture.image.width,
+    height: dataTexture.image.height,
+    type: UnsignedByteType,
+    colorSpace: LinearSRGBColorSpace,
+    material,
+    renderer,
+    renderTargetOptions: params.renderTargetOptions
+  })
   try {
     quadRenderer.render()
   } catch (e) {
-    quadRenderer.dispose()
+    quadRenderer.disposeOnDemandRenderer()
     throw e
   }
   return quadRenderer